Add basic HTTP endpoint for querying via JSON
This more fully fleshes out the logic from the previous commit. I did
have some help from an LLM. At the bottom of this commit message is the
initial prompt I used.

The code was fine enough. I did a lot of back and forth and cleaning up.

---

We are writing a Flask application meant to explore embeddings created from a knowledge graph. Each embedding in that knowledge graph represents an RDF node in that graph. Here's an example of one of those nodes, written as the JSON representation of a point from Qdrant:

```
{"id":1,"payload":{"graph":"nasa-gesdisc","iri":"https://nasa-gesdisc.proto-okn.net/kg/node/0","label":"ERS-1_BYU_L3_OW_SIGMA0_ENHANCED; subject: ERS-1 Gridded Level 3 Enhanced Resolution Sigma-0 from BYU"},"vector":[ ...a vector created from the text representation in `label`... ]}
```

These embeddings have already been created in a different repository. They were created using `SentenceTransformer('all-MiniLM-L6-v2')`.

What we want to do is to build an HTTP endpoint where we can start to query these embeddings based on similarity. We want to allow searching for similarity in two different ways:

A. Providing a URI and showing nodes similar to that one

B. Providing some text, converting that into a vector using the same transformer, and showing nodes similar to that one

First things first, I want to create an endpoint, let's call it `/query`, where we can provide one or more of the parameters in A or B, then query Qdrant for similar points. So, we can say "I want nodes similar to the node http://example.org/A and also to the text 'hydrology'". Let's make this receive a POST that accepts JSON. I've started a mockup of this query object in pydantic in `model.py`. Note that I mocked up executing this query in an incomplete way in `query.py`. Note also that I made it so you can give it *negative* parameters as well, as in "but also penalize nodes that have the text 'united states'".

Go ahead and write a simple Flask application that does this. You can use my code in `query.py` as a guide. If it seems correct, please use it as much as possible. Note that some of the functions include an unbound `client` variable. This is meant to be the Qdrant client. I have not written flask in a while, and I didn't remember exactly how to include globals like this (which is basically the same as maintaining a database connection). I'm not connected to the qdrant server right now, so don't run anything. We will get there.